+++ scout/server/blueprints/institutes/forms.py
@@ -13,10 +13,9 @@ from wtforms import (
 )
 from wtforms.widgets import PasswordInput, TextInput
 
-from scout.constants import CASE_SEARCH_TERMS, PHENOTYPE_GROUPS
+from scout.constants import PHENOTYPE_GROUPS
 from scout.models.case import STATUS
 
-CASE_SEARCH_KEY = [(value["prefix"], value["label"]) for key, value in CASE_SEARCH_TERMS.items()]
 CATEGORY_CHOICES = [("snv", "SNV"), ("sv", "SV")]