+++ tests/server/blueprints/dashboard/test_dashboard_controllers.py
@@ -1,13 +1,13 @@
 from flask import url_for
+from werkzeug.datastructures import ImmutableMultiDict
 
 from scout.server.blueprints.dashboard.controllers import (
-    compose_slice_query,
     dashboard_form,
     get_dashboard_info,
     institute_select_choices,
 )
 from scout.server.blueprints.dashboard.forms import DashboardFilterForm
-from scout.server.extensions import store
+from scout.server.blueprints.institutes.forms import CaseFilterForm
 
 
 def test_institute_select_choices(user_obj, app):
@@ -37,20 +37,8 @@ def test_dashboard_form(app):
         # A DashboardFilterForm should be created correctly by dashboard_form
         df = dashboard_form(None)
         assert isinstance(df, DashboardFilterForm)
-        assert df.search_type
+        assert isinstance(df, CaseFilterForm)
         assert df.search_institute
-        assert df.search_term
-        assert df.search
-
-
-def test_compose_slice_query(app):
-    """Test function that combines form fields to create filter query"""
-
-    # GIVEN two parameters provided in the dashboard filter form
-    search_type = "case:"
-    search_term = "test_case"
-    slice_query = compose_slice_query(search_type, search_term)
-    assert slice_query == f"{search_type}{search_term}"
 
 
 def test_empty_database(real_adapter):
@@ -95,7 +83,7 @@ def test_one_causative(real_adapter, case_obj):
             assert info["count"] == 0
 
 
-def test_with_slice_query(real_adapter, case_obj):
+def test_with_name_query(real_adapter, case_obj):
     ## GIVEN an database with one case
     adapter = real_adapter
     adapter.case_collection.insert_one(case_obj)
@@ -104,8 +92,12 @@ def test_with_slice_query(real_adapter, case_obj):
 
     institute_id = case_obj["owner"]
 
-    slice_query = f"case:{case_display_id}"
-    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
+    name_query = ImmutableMultiDict(
+        {
+            "case": case_display_id,
+        }
+    )
+    data = get_dashboard_info(adapter, institute_id=institute_id, cases_form=name_query)
 
     ## THEN assert there is one case in the data
     for group in data["cases"]:
@@ -128,9 +120,13 @@ def test_with_hpo_query(real_adapter, case_obj):
 
     ## WHEN querying for cases with that phenotype id
     institute_id = case_obj["owner"]
-    slice_query = f"exact_pheno:{phenotype['phenotype_id']}"
+    name_query = ImmutableMultiDict(
+        {
+            "exact_pheno": phenotype["phenotype_id"],
+        }
+    )
 
-    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
+    data = get_dashboard_info(adapter, institute_id=institute_id, cases_form=name_query)
     ## THEN assert there is one case in the data
     for group in data["cases"]:
         if group["status"] == "all":
@@ -152,9 +148,13 @@ def test_with_phenotype_group_query(real_adapter, case_obj):
 
     ## WHEN querying for cases with that phenotype id
     institute_id = case_obj["owner"]
-    slice_query = "pheno_group:HP:0000001"
+    name_query = ImmutableMultiDict(
+        {
+            "pheno_group": "HP:0000001",
+        }
+    )
 
-    data = get_dashboard_info(adapter, institute_id=institute_id, slice_query=slice_query)
+    data = get_dashboard_info(adapter, institute_id=institute_id, cases_form=name_query)
     ## THEN assert there is one case in the data
     for group in data["cases"]:
         if group["status"] == "all":